add a new function module and move functions into it
