+++ pypose/basics/__init__.py
@@ -1,2 +1,2 @@
-from .function import cart2homo, homo2cart
+# functions implemented in this module are needed by lietensor module
 from .ops import pm, cumops, cummul, cumprod, cumops_, cummul_, cumprod_
