+++ pypose/function/__init__.py
@@ -1,2 +1,2 @@
-
+# functions implemented in this module need the lietensor module
 from .geometry import *
